Fixes for dev httr2
httr2 has switched to a more aggressive URL parser which now throws an error, so I've made the minimal fixes to get your tests passing again. In this case, that was removing a test for an undocumented feature.

The proposed code should work with both old and new httr2, so you can submit to CRAN now, if you want. I am planning to submit httr2 to CRAN on Jan 17, so you'll have to submit ~2 weeks after this date.